+++ R/plot_histogram.R
@@ -76,7 +76,7 @@ plot_histogram <- S7::new_generic(
 #' )
 #'
 #' @name plot_histogram_moo
-#' @method plot_histogram multiOmicDataSet
+#' @usage NULL
 #' @seealso [plot_histogram] generic
 #' @family plotters for multiOmicDataSets
 S7::method(plot_histogram, multiOmicDataSet) <- function(
@@ -150,6 +150,7 @@ S7::method(plot_histogram, multiOmicDataSet) <- function(
 #'
 #' @seealso [plot_histogram] generic
 #' @name plot_histogram_dat
+#' @usage NULL
 #' @method plot_histogram data.frame
 #' @family plotters for counts dataframes
 S7::method(plot_histogram, S7::class_data.frame) <- function(